Move event sorting into a standalone function
