+++ content/webapp/components/EventsByMonth/EventsByMonth.tsx
@@ -1,6 +1,4 @@
 import { Component } from 'react';
-import sortBy from 'lodash.sortby';
-import { getEarliestFutureDateRange } from '@weco/common/utils/dates';
 import { classNames, cssGrid } from '@weco/common/utils/classnames';
 import SegmentedControl from '@weco/common/views/components/SegmentedControl/SegmentedControl';
 import { EventBasic } from '../../types/events';
@@ -64,18 +62,10 @@ class EventsByMonth extends Component<Props, State> {
 
     // Need to order the events for each month based on their earliest future date range
     Object.keys(eventsInMonths).map(label => {
-      eventsInMonths[label] = sortBy(eventsInMonths[label], [
-        m => {
-          const times = m.times.map(time => ({
-            start: time.range.startDateTime,
-            end: time.range.endDateTime,
-          }));
-          const fromDate = startOf(parseLabel(label));
-
-          const earliestRange = getEarliestFutureDateRange(times, fromDate);
-          return earliestRange && earliestRange.start;
-        },
-      ]);
+      eventsInMonths[label] = sortByEarliestFutureDateRange(
+        label,
+        eventsInMonths[label]
+      );
     });
 
     return (
